feat: :sparkles: Fix leaks, diameter, color range
릭 체크, 지름->반지름, 컬러 범위 255->1로 조정

#15
