+++ cjang.rt
@@ -1,9 +1,11 @@
 A 0.2                  255,255,255
 
 
-C -50,0,20	 	0,0,1		70
-L -40,0,30		    	      			0.7			        255,254,0
+C -2,2,1	 	-1,1,1	90
+L 0,20,0		    	      			0.6			        255,255,255
 
-pl 0,0,0	    		   		0,1.0,0								255,0,225
-sp 0,0,20	    									20 					255,0,0
-cy 50.0,0.0,20.6	0,0,1.0		14.2 21.42	10,0,255
+sp 0,-100.5,-1	    									200 					204,204,0
+sp 0,0,-1  1  25,51,127
+sp -1,0,-1  1  204,204,25
+sp 1,0,-1  1  204,153,51
+
