+++ src/main.c
@@ -1,5 +1,6 @@
 #include "mlx.h"
 #include "parse.h"
+#include "list.h"
 #include "x11_events.h"
 #include "event.h"
 
@@ -20,11 +21,12 @@ int	main(int argv, char **argc)
 	t_mlx		mlx;
 
 	fd = fd_get(argv, argc);
-	info = info_get(fd);
+	info = parse_to_str(fd);
 	scene = parse_to_obj(info);
+	del_obj_lst_parse(&info);
 
 	program_init(&mlx);
-	scene_draw(&mlx);
+	scene_draw(&mlx, scene);
 
 	mlx_hook(mlx.win, X11_KEYPRESS, 1L << 0, key_press, &mlx);
 	mlx_hook(mlx.win, X11_CLOSEBTN, 1L << 2, exit_button, &mlx);
