+++ src/parse/parse_to_str.c
@@ -3,47 +3,8 @@
 #include "vector3.h"
 #include "list.h"
 #include "error.h"
+#include <stdlib.h>
 
-int	is_range(int value_min, int value_max, int value_cur)
-{
-	if (value_cur < value_min || value_cur > value_max)
-		return (0);
-	return (1);
-}
-
-// enum에 맞게 고쳐야할 지도 모름
-// is_???? 로 바꾸기
-int	obj_cmp(char *s)
-{
-	if (ft_strcmp(s, "A") && ft_strcmp(s, "C") && ft_strcmp(s, "L") \
-	&& ft_strcmp(s, "sp") && ft_strcmp(s, "pl") && ft_strcmp(s, "cy"))
-		return (0);
-	return (1);
-}
-
-t_parse	*new_parse(void)
-{
-	t_parse		*lst;
-
-	lst = ft_calloc(sizeof(t_parse), 0);
-	return (lst);
-}
-
-/*
-t_bool로 처리하는게 좋을듯.
-참조했던 정보들의 갯수 != 들어온 정보의 개수 -> 인자가 잘못 들어온 경우임
-*/
-int		is_element_valid(char **str, int idx)
-{
-	int		i;
-
-	i = 0;
-	while (str[i] != NULL)
-		++i;
-	if (i != idx)
-		return (0);
-	return (1);
-}
 
 // 25줄 수정요망
 void	parse_set(t_parse *lst, char **str)
@@ -51,7 +12,7 @@ void	parse_set(t_parse *lst, char **str)
 	char	*id;
 	int		idx;
 
-	if (obj_cmp(str[0]) == 0)
+	if (is_object(str[0]) == 0)
 		error_user("invalid object name.\n");
 	else
 		lst->ident = str[0];
@@ -78,50 +39,6 @@ void	parse_set(t_parse *lst, char **str)
 	return ;
 }
 
-static t_parse	*split_free(char **str)
-{
-	int		idx;
-
-	idx = 0;
-	while (str[idx] != NULL)
-		free(str[idx++]);
-	free(str);
-	return (NULL);
-}
-
-void	parse_print(t_obj_list *lst)
-{
-	t_parse		*ps;
-	char		*id;
-
-	if (lst == NULL)
-		return ;
-	while (lst)
-	{
-		ps = lst->object;
-		id = ps->ident;
-		printf("%s\t", id);
-		if (!ft_strcmp(id, "C") || !ft_strcmp(id, "L") || \
-		!ft_strcmp(id, "sp") || !ft_strcmp(id, "pl") || !ft_strcmp(id, "cy"))
-			printf("%s\t", ps->point);
-		if (!ft_strcmp(id, "A") || !ft_strcmp(id, "L"))
-			printf("%s\t", ps->bri_ratio);
-		if (!ft_strcmp(id, "C") || !ft_strcmp(id, "pl") || !ft_strcmp(id, "cy"))
-			printf("%s\t", ps->nor_vec);
-		if (!ft_strcmp(id, "sp") || !ft_strcmp(id, "cy"))
-			printf("%s\t", ps->diameter);
-		if (!ft_strcmp(id, "cy"))
-			printf("%s\t", ps->height);
-		if (!ft_strcmp(id, "C"))
-			printf("%s\t", ps->fov);
-		if (!ft_strcmp(id, "A") || !(ft_strcmp(id, "L")) || \
-		!ft_strcmp(id, "sp") || !ft_strcmp(id, "pl") || !ft_strcmp(id, "cy"))
-			printf("%s\t", ps->rgb);
-		printf("\n");
-		lst = lst->next;
-	}
-}
-
 t_parse	*info_set(char *line)
 {
 	char		**str_splited;
@@ -132,7 +49,10 @@ t_parse	*info_set(char *line)
 	if (str_splited == NULL)
 		error_user("info_set - dynamic allocation problem.\n");
 	else if (str_splited[0] == NULL)
-		return (split_free(str_splited));
+	{
+		free(lst);
+		return (del_split(str_splited));
+	}
 	parse_set(lst, str_splited);
 	free(str_splited);
 	return (lst);
@@ -140,7 +60,7 @@ t_parse	*info_set(char *line)
 
 // (x,y,z)가 오는 경우 쉼표 사이에 띄어쓰기 가능??
 // 대문자가 하나도 없거나 여러개에 대한 대비
-t_obj_list	*info_get(int fd)
+t_obj_list	*parse_to_str(int fd)
 {
 	char		*line;
 	int			gnl_ret;
@@ -163,5 +83,6 @@ t_obj_list	*info_get(int fd)
 			free(line);
 	}
 	// parse_print(lst_head);
+
 	return (lst_head);
 }
